\name{Gammareg-package}
\alias{Gammareg-package}
\docType{package}
\title{classic gamma regression: joint modeling of mean and shape parameters}
\description{Classic gamma regression package}
\details{
\tabular{ll}{
Package: \tab Gammareg\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-01-23\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
 Martha Corrales and Edilberto Cepeda-Cuervo with the colaboration of Maria Fernanda Zarate,  
 Ricardo Duplat and  Campo Elias Pardo.
}
\keyword{package}




