\name{print.summary.Gammareg}
\alias{print.summary.Gammareg}
\title{print the summary of the Classic gamma regression}
\description{Print the summary Classic Gamma Regression for joint modelling of mean and shape parameters.}
\usage{
\method{print}{summary.Gammareg}(x, \ldots)
}
\arguments{
 \item{x}{object of class Gammareg}
 \item{\ldots}{not used.}
}
\value{
Print the summary Classic Gamma Regression for joint modelling of mean and shape parameters.
}
\references{
1. Cepeda-Cuervo E. (2001). Modelagem da variabilidade em modelos lineares generalizados. Unpublished Ph.D. tesis. Instituto de Matemáticas. 
Universidade Federal do Río do Janeiro.
//http://www.docentes.unal.edu.co/ecepedac/docs/MODELAGEM20DA20VARIABILIDADE.pdf.
http://www.bdigital.unal.edu.co/9394/.
2. McCullagh, P. and Nelder, N.A. (1989). Generalized Linear Models. Second Edition. Chapman and Hall. 
}
\author{
 Martha Corrales \email{martha.corrales@usa.edu.co}      
 Edilberto Cepeda-Cuervo \email{ecepedac@unal.edu.co}        
}
\keyword{Classic estimation}
\keyword{Fisher Scoring}
\keyword{joint modeling}
\keyword{Gamma regression}
\keyword{Mean and shape parameters}